Add files via upload
File 1: Database Used in the Project
This file serves as the database utilized within the project. It contains altered values to safeguard privacy concerns while maintaining the integrity of the data. The database primarily stores information regarding suppliers and our purchasing activities. Due to privacy constraints, certain values have been modified to ensure anonymity and confidentiality.

File 2: PDF Document Related to the Power BI File
This PDF document corresponds to the Power BI dashboard created for the project. It provides a visual representation of the data extracted from the database. However, due to privacy considerations, the Power BI file itself cannot be shared. The PDF serves as a static representation of the insights derived from the dashboard without compromising sensitive information.